+++ manager/api/tests/test_ole.py
@@ -28,6 +28,8 @@ from django.test import TestCase, override_settings
 from django.urls import reverse
 from rest_framework.test import APIClient
 
+from manager.utils import DATETIME_ISO_FORMAT, get_tai_from_utc
+
 
 @override_settings(DEBUG=True)
 class OLETestCase(TestCase):
@@ -71,9 +73,9 @@ class OLETestCase(TestCase):
             "components": "MainTel",
             "primary_software_components": "None",
             "primary_hardware_components": "None",
-            "date_begin": "202200703-19:58:13",
-            "date_end": "20220704-19:25:13",
-            "time_lost": 10,
+            "date_begin": "2024-01-01T00:00:00.000000",
+            "date_end": "2024-01-01T00:10:00.000000",
+            "time_lost": 1,
             "level": 0,
         }
 
@@ -139,6 +141,8 @@ class OLETestCase(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.data), 0)
 
+        mock_ole_client.stop()
+
     def test_simple_exposurelog_create(self):
         """Test exposurelog create."""
         # Arrange:
@@ -158,6 +162,8 @@ class OLETestCase(TestCase):
         response = self.client.post(url, self.payload_full_exposure)
         self.assertEqual(response.status_code, 201)
 
+        mock_ole_client.stop()
+
     def test_exposurelog_update(self):
         """Test exposurelog update."""
         # Arrange:
@@ -176,6 +182,8 @@ class OLETestCase(TestCase):
         url = reverse("ExposureLogs-detail", args=[1])
         response = self.client.put(url, self.payload_full_exposure)
         self.assertEqual(response.status_code, 200)
+
+        mock_ole_client.stop()
 
     def test_exposurelog_create_with_jira(self):
         """Test exposurelog create with jira."""
@@ -222,6 +230,9 @@ class OLETestCase(TestCase):
         response = self.client.post(url, self.payload_full_exposure_with_jira_comment)
         self.assertEqual(response.status_code, 201)
 
+        mock_jira_ticket_client.stop()
+        mock_ole_client.stop()
+
     def test_exposurelog_update_with_jira(self):
         """Test exposurelog update with jira."""
         # Arrange:
@@ -267,6 +278,9 @@ class OLETestCase(TestCase):
         response = self.client.put(url, self.payload_full_exposure_with_jira_comment)
         self.assertEqual(response.status_code, 200)
 
+        mock_jira_ticket_client.stop()
+        mock_ole_client.stop()
+
     def test_narrativelog_list(self):
         """Test narrativelog list."""
         # Arrange:
@@ -287,6 +301,8 @@ class OLETestCase(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.data), 0)
 
+        mock_ole_client.stop()
+
     def test_simple_narrativelog_create(self):
         """Test narrativelog create."""
         # Arrange:
@@ -304,7 +320,25 @@ class OLETestCase(TestCase):
         # Act:
         url = reverse("NarrativeLogs-list")
         response = self.client.post(url, self.payload_full_narrative)
+
+        # Assert:
         self.assertEqual(response.status_code, 201)
+
+        # Check the date_begin and date_end arguments
+        # are transformed to TAI scale.
+        mock_ole_client_json_arg = mock_ole_client.call_args.kwargs["json"].dict()
+        date_begin_arg = mock_ole_client_json_arg["date_begin"]
+        date_end_arg = mock_ole_client_json_arg["date_end"]
+        payload_date_begin_formatted = get_tai_from_utc(
+            self.payload_full_narrative["date_begin"]
+        ).strftime("%Y-%m-%dT%H:%M:%S.%f")
+        payload_date_end_formatted = get_tai_from_utc(
+            self.payload_full_narrative["date_end"]
+        ).strftime("%Y-%m-%dT%H:%M:%S.%f")
+        assert date_begin_arg == payload_date_begin_formatted
+        assert date_end_arg == payload_date_end_formatted
+
+        mock_ole_client.stop()
 
     def test_narrativelog_update(self):
         """Test narrativelog update."""
@@ -323,7 +357,25 @@ class OLETestCase(TestCase):
         # Act:
         url = reverse("NarrativeLogs-detail", args=[1])
         response = self.client.put(url, self.payload_full_narrative)
+
+        # Assert
         self.assertEqual(response.status_code, 200)
+
+        # Check the date_begin and date_end arguments
+        # are transformed to TAI scale.
+        mock_ole_client_json_arg = mock_ole_client.call_args.kwargs["json"].dict()
+        date_begin_arg = mock_ole_client_json_arg["date_begin"]
+        date_end_arg = mock_ole_client_json_arg["date_end"]
+        payload_date_begin_formatted = get_tai_from_utc(
+            self.payload_full_narrative["date_begin"]
+        ).strftime(DATETIME_ISO_FORMAT)
+        payload_date_end_formatted = get_tai_from_utc(
+            self.payload_full_narrative["date_end"]
+        ).strftime(DATETIME_ISO_FORMAT)
+        assert date_begin_arg == payload_date_begin_formatted
+        assert date_end_arg == payload_date_end_formatted
+
+        mock_ole_client.stop()
 
     def test_narrative_log_create_with_jira(self):
         """Test narrativelog create with jira."""
@@ -370,6 +422,9 @@ class OLETestCase(TestCase):
         response = self.client.post(url, self.payload_full_narrative_with_jira_comment)
         self.assertEqual(response.status_code, 201)
 
+        mock_jira_ticket_client.stop()
+        mock_ole_client.stop()
+
     def test_narrative_log_update_with_jira(self):
         """Test narrativelog update with jira."""
         # Arrange:
@@ -414,6 +469,9 @@ class OLETestCase(TestCase):
         url = reverse("NarrativeLogs-detail", args=[1])
         response = self.client.put(url, self.payload_full_narrative_with_jira_comment)
         self.assertEqual(response.status_code, 200)
+
+        mock_jira_ticket_client.stop()
+        mock_ole_client.stop()
 
 
 @override_settings(DEBUG=True)
@@ -485,6 +543,8 @@ class NightReportTestCase(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.data), 1)
 
+        mock_ole_client.stop()
+
     def test_simple_nightreport_create(self):
         """Test nightreport create."""
         # Arrange:
@@ -503,6 +563,8 @@ class NightReportTestCase(TestCase):
         url = reverse("NightReportLogs-list")
         response = self.client.post(url, self.payload)
         self.assertEqual(response.status_code, 201)
+
+        mock_ole_client.stop()
 
     def test_nightreport_update(self):
         """Test nightreport update."""
@@ -523,6 +585,8 @@ class NightReportTestCase(TestCase):
         response = self.client.put(url, self.payload)
         self.assertEqual(response.status_code, 200)
 
+        mock_ole_client.stop()
+
     def test_nightreport_delete(self):
         """Test nightreport delete."""
         # Arrange:
@@ -541,6 +605,8 @@ class NightReportTestCase(TestCase):
         url = reverse("NightReportLogs-detail", args=[1])
         response = self.client.delete(url)
         self.assertEqual(response.status_code, 200)
+
+        mock_ole_client.stop()
 
     def test_nightreport_send(self):
         """Test nightreport send."""
@@ -578,6 +644,11 @@ class NightReportTestCase(TestCase):
         response = self.client.post(url)
         self.assertEqual(response.status_code, 200)
 
+        mock_ole_client_get.stop()
+        mock_ole_client_patch.stop()
+        mock_get_jira_obs_report_client.stop()
+        mock_send_smtp_email_client.stop()
+
     def test_nightreport_already_sent(self):
         """Test nightreport already sent."""
         # Arrange:
@@ -601,3 +672,5 @@ class NightReportTestCase(TestCase):
         response = self.client.post(url)
         self.assertEqual(response.status_code, 400)
         self.assertEqual(response.data, {"error": "Night report already sent"})
+
+        mock_ole_client_get.stop()